* sdcc/support/regression/tests/bug-460010.c: fix seg violation on host
	* sdcc/support/regression/tests/muldiv.c: fix output on mcs51

	* sdcc/support/regression/tests/stacks.c: fix DSEG overflow on mcs51


git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@1613 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ support/regression/tests/muldiv.c
@@ -35,7 +35,7 @@ testUnsignedMul(void)
 
     i = 37;
 
-    LOG(("i*3 == 111 = %u\n", i*3));
+    LOG(("i*3 == 111 = %u\n", (int)(i*3)));
     result = i*3;
     ASSERT(result == 111);
